make KafkaStreamTarget into a class
